+++ dwds/CHANGELOG.md
@@ -9,6 +9,9 @@
 - Compute only the required top frame for a paused event. 
 - Change `streamListen` to return an `RPCError` / error code `-32601` for streams
   that are not handled.
+- Populate information about async Dart frames.
+- Populate the `exception` field in debugger pause event when we break as a result
+  of an exception.
 - Prompt users to install the Dart Debug Extension if local debugging does not work.
 - Allow for the injected client to run with CSP enforced.
 - Implement the `getMemoryUsage()` call.